1.updated the function call of the functionbearer 2.destroy function is updated for the authentication and authorization 3.updated the time of the expreiesIn token 4.router is created for the delete the post using API

+++ controllers/api/v1/posts_api.js
@@ -21,16 +21,20 @@ module.exports.destroy = async (req, res) => {
   try {
     let post = await Post.findById(req.params.id);
 
-    // if (post.user == req.user.id) {
-    await Post.findByIdAndDelete(req.params.id);
-    await Comment.deleteMany({ post: req.params.id });
-    // req.flash("success", "Post Deleted Successfully!");
-    // return res.redirect("back");
-    // }
-    return res.status(200).json({
-      message: "Post Deleted Successfully",
-      post: post,
-    });
+    if (post.user == req.user.id) {
+      await Post.findByIdAndDelete(req.params.id);
+      await Comment.deleteMany({ post: req.params.id });
+      // req.flash("success", "Post Deleted Successfully!");
+      // return res.redirect("back");
+      return res.status(200).json({
+        message: "Post Deleted Successfully",
+        post: post,
+      });
+    } else {
+      return res.status(401).json({
+        message: "You can't delete this post!",
+      });
+    }
   } catch (err) {
     // console.log("Error while deleting Post: " + err);
     // return res.redirect("back");
